refactor(py_src): split per-language helpers into emit/{elixir,erlang}/
helper.py held five pairs of parallel `_elixir` / `_erlang` functions
(when_guard, map_argname, map_argtype_to_guard, map_argtype_in_docs,
map_argtype_in_spec) plus `map_uppercase_to_erlang_name`. Same logic
twice, mutual independence — exactly the wrong place for the kitchen-sink
helper module.

  py_src/emit/__init__.py
  py_src/emit/elixir/__init__.py
  py_src/emit/elixir/helpers.py  (220 LoC: 5 *_elixir functions)
  py_src/emit/erlang/__init__.py
  py_src/emit/erlang/helpers.py  (272 LoC: 5 *_erlang functions
                                  + map_uppercase_to_erlang_name)

helper.py keeps the five dispatchers (`map_argtype_in_spec(kind, ...)`
etc.) as thin shims — they lazy-import the per-language implementations
to break the `helper -> emit -> helper` cycle that would otherwise form
because the emit modules `from helper import *` to reach
`reserved_keywords`, `is_int_type`, `is_struct`, `vec_out_types`, etc.

helper.py: 1092 -> 631 LoC (-461).

Adding a third BEAM target now means dropping a sibling under emit/
rather than threading a third `elif kind == 'X':` through five places.

Byte-equivalent vs lib-4.11.0/ and src-4.11.0/.
